requirements.txt
mediapipe
opencv-python
numpy
streamlit
streamlit-lottie
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
matplotlib
plotly



